pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "pdf-to-markdown"
version = "0.1.0"
description = "Convert PDF documents to Markdown with embedded image extraction and optional OCR"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]

[project.scripts]
pdf-to-markdown = "pdf_to_markdown.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
